rename PG_DC
