+++ qulab/drivers/PG_DC/DC_driver.py
@@ -3,11 +3,11 @@ import numpy as np
 
 from qulab import BaseDriver, QInteger, QOption, QReal, QString, QVector
 
-from .Highfine_DCsource import Voltage
+from .PG_DC_api import Voltage
 
 
 class Driver(BaseDriver):
-    support_models = ['Highfine_DCsource']
+    support_models = ['PG_DC']
     quants = [
         QReal('Offset', value=0, ch=0),
             ]
@@ -15,6 +15,7 @@ class Driver(BaseDriver):
     def __init__(self, **kw):
         BaseDriver.__init__(self, **kw)
         self.ip=kw['IP']
+        self.model = 'PG_DC'
 
     def performOpen(self):
         self.handle = Voltage(self.ip)
